# Conflicts:
#	app/models/Photo.java
#	app/repository/PhotoRepository.java
#	conf/test/country/evolutions/default/1.sql
#	conf/test/destination/evolutions/default/1.sql
#	conf/test/photo/evolutions/default/1.sql
#	conf/test/photo/evolutions/default/2.sql
#	conf/test/profile/evolutions/default/1.sql
#	conf/test/treasureHunt/evolutions/default/1.sql
#	conf/test/trip/evolutions/default/1.sql
#	conf/test/user/evolutions/default/1.sql

+++ app/repository/PhotoRepository.java
@@ -9,14 +9,12 @@ import io.ebean.Ebean;
 import io.ebean.EbeanServer;
 import java.util.Collection;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import models.DestinationPhoto;
 import models.Photo;
 import play.db.ebean.EbeanConfig;
-import play.mvc.Result;
 import util.objects.Pair;
 
 /**
@@ -25,7 +23,7 @@ import util.objects.Pair;
 @Singleton
 public class PhotoRepository {
 
-    public static final String FRONTEND_APPEND_DIRECTORY = "../user_content/";
+    private static final String FRONTEND_APPEND_DIRECTORY = "../user_content/";
     private static final String USER_ID = "user_id";
     private final EbeanServer ebeanServer;
     private final DatabaseExecutionContext executionContext;
@@ -42,15 +40,40 @@ public class PhotoRepository {
      * @param photo Photo to be added
      * @return Ok on success
      */
-    public CompletableFuture<Result> addPhoto(Photo photo) {
+    public CompletableFuture<Long> addPhoto(Photo photo) {
         return supplyAsync(() -> {
             ebeanServer.insert(photo);
-            return ok();
+            return photo.guid;
         }, executionContext);
     }
 
     /**
-     * Finds all photos in database related to the given user ID.
+     * Clears any existing profile photo for a user, and returns filenames of old files.
+     *
+     * @param userID user ID to clear profile photo of
+     * @return A Pair of strings, the key being the filename of the photo and the value being the
+     * thumbnail filename of the photo. Returns null if the user didn't have a profile photo
+     */
+    public CompletableFuture<Pair<String, String>> clearProfilePhoto(long userID) {
+        return supplyAsync(() -> {
+            Pair<String, String> returnPair;
+            Photo profilePhoto = ebeanServer.find(Photo.class)
+                .where()
+                .eq(USER_ID, userID)
+                .eq(IS_PROFILE, true)
+                .findOneOrEmpty().orElse(null);
+            if (profilePhoto == null) {
+                return null;
+            } else {
+                returnPair = new Pair<>(profilePhoto.filename, profilePhoto.thumbnailFilename);
+                ebeanServer.delete(profilePhoto);
+                return returnPair;
+            }
+        });
+    }
+
+    /**
+     * Finds all photos in database related to the given user ID that aren't a profile photo
      *
      * @param userID User to find all photos for
      * @return List of Photo objects with the specified user ID
@@ -68,7 +91,7 @@ public class PhotoRepository {
     }
 
     /**
-     * Finds all public photos in database related to the given user ID.
+     * Finds all public photos in database related to the given user ID that aren't a profile photo
      *
      * @param userID User to find all public photos for
      * @return List of Photo objects with the specified user ID
@@ -87,6 +110,30 @@ public class PhotoRepository {
     }
 
     /**
+     * Finds the profile picture in the database for the given user ID.
+     *
+     * @param userID User to find profile picture for
+     * @return a photo, which will be null if they user doesn't exist or doesn't have a profile
+     * picture
+     */
+    public CompletableFuture<Photo> getUserProfilePicture(Long userID) {
+        return supplyAsync(() -> {
+            Photo photo = ebeanServer.find(Photo.class)
+                .where()
+                .eq(USER_ID, userID)
+                .eq(IS_PROFILE, true)
+                .findOneOrEmpty().orElse(null);
+
+            if (photo != null) {
+                photo.filename = FRONTEND_APPEND_DIRECTORY + photo.filename;
+                photo.thumbnailFilename = FRONTEND_APPEND_DIRECTORY + photo.thumbnailFilename;
+            }
+
+            return photo;
+        }, executionContext);
+    }
+
+    /**
      * Adds photos into the database. Will replace the users profile picture if needed.
      *
      * @param photos A list of photos to upload
@@ -99,7 +146,7 @@ public class PhotoRepository {
      * Deletes a photo from the database.
      *
      * @param id Unique photo ID of destination to be deleted
-     * @return The number of rows deleted
+     * @return The deleted photo, or null if a photo with that id was not found
      */
     public CompletableFuture<Photo> deletePhoto(Long id) {
         return supplyAsync(() -> {
@@ -109,7 +156,9 @@ public class PhotoRepository {
                     .findOneOrEmpty()
                     .orElse(null);
 
-                photo.delete();
+                if (photo != null) {
+                    photo.delete();
+                }
                 return photo;
             }
             , executionContext);
@@ -146,21 +195,7 @@ public class PhotoRepository {
      * Get photo object from db where it has some filename.
      *
      * @param filename Name of file
-     */
-    public CompletableFuture<Photo> getPhotoByFilename(String filename) {
-        return supplyAsync(() -> ebeanServer.find(Photo.class)
-                .where()
-                .eq("filename", filename)
-                .findOneOrEmpty()
-                .orElse(null),
-            executionContext);
-
-    }
-
-    /**
-     * Get photo object from db where it has some filename.
-     *
-     * @param filename Name of file
+     * @return true if deleted, false if not
      */
     public CompletableFuture<Boolean> deletePhotoByFilename(String filename) {
         return supplyAsync(() -> {
@@ -168,7 +203,7 @@ public class PhotoRepository {
                 .where()
                 .eq("filename", filename)
                 .findOne();
-            if(photo != null) {
+            if (photo != null) {
                 photo.delete();
                 return true;
             } else {
@@ -181,6 +216,7 @@ public class PhotoRepository {
      * Update photo row in database.
      *
      * @param photo photo object to update
+     * @return the updated photo's guid
      */
     public CompletableFuture<Long> updatePhoto(Photo photo) {
         return supplyAsync(() -> {
@@ -190,6 +226,13 @@ public class PhotoRepository {
             executionContext);
     }
 
+    /**
+     * Gets a deleted destination photo form the database
+     *
+     * @param photoId The id of the photo to retrieve
+     * @param destId The id of the associated destination
+     * @return The destination photo, or null if not found
+     */
     public CompletableFuture<DestinationPhoto> getDeletedDestPhoto(Long photoId, Long destId) {
         return supplyAsync(() -> ebeanServer.find(DestinationPhoto.class)
             .setIncludeSoftDeletes()
@@ -198,5 +241,5 @@ public class PhotoRepository {
             .eq("destination_id", destId)
             .findOneOrEmpty()
             .orElse(null), executionContext);
-    };
+    }
 }
